requirements.txt
chris_plugin==0.0.17
pycivet==0.0.2a1
loguru~=0.6.0
